+++ CHANGELOG.md
@@ -5,6 +5,12 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Add `port/protocol` to `TrafficFilterRules` to filter traffic in Piglet
+
 ## [0.5.0] - 2023-04-24
 
 ### Changed